data-ingestion/requirements.txt
# Alpha Vantage Data Ingestion Requirements
# ==========================================

# Core dependencies
requests>=2.31.0
pandas>=2.0.0
boto3>=1.28.0
botocore>=1.31.0

# Data processing
numpy>=1.24.0
pytz>=2023.3

# API rate limiting
ratelimit>=2.2.1

# Logging and monitoring
structlog>=23.1.0

# Configuration management
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.1
moto>=4.2.0  # For AWS mocking

# Type hints
typing-extensions>=4.7.0

# Data validation
marshmallow>=3.20.0

# Async support (pour futures améliorations)
aiohttp>=3.8.5
asyncio-throttle>=1.0.2

# Retry mechanisms
tenacity>=8.2.3

# Time handling
arrow>=1.2.3

# JSON handling
orjson>=3.9.0

# Environment detection
python-environ>=0.4.54
